Add files via upload
There, just need to create the logic flow for the types of will button and the on or before date to address the legislative change. This version adds in the rest of the types of clauses for different capacity levels and fixed the spouse as witness no typo causing it not to work.