Add files via upload
I declared a global variable for the first switch statement, and removed 'let' from inside the function with the switch statement, and that made the values accessible to subsequent functions (made them globally accessible); You need to add in the none of the above else if branch for that radio option for the how's it written/handwritten function addItem7 still. Then you'll have a working example so that the next ones are just cut and paste and replace.