Add files via upload
There, the bare basic logic is in except for statutory will questions. Next is confirm everything, add in statutes and links, and the formatting function to sort the responses and get rid of duplicates. Then you can start going into detail on if the content of the will is valid if you feel like it.