Add files via upload
I added in for the type of will -- to capacity -- and that if olographic then you don't need a testator clause anyway. Okay, so you need to get in the exact dates for the statutory will part now, I think it will be from the one citation in the one link you already have in about nuncupatives, and then I think it ends when the notarial law was enacted, except one is really a clarification on the other. As far as if you need testator clauses for mystic and nuncupative wills, that is something I need to think on and study more, cuz i never saw one of those wills at Acadiana and need to learn that. I will bone up and read on it.