Add files via upload
In this version you added the show/hide function for the section detailing the archaic will types. You were messing with customizing the button to make it pretty with a custom rounded border when it was time to go to work. Do that if you still want to next time you dip in. Also check if the show hide button will work on a touch screen. I think it will ... Now is coding in the show/hide logic flow for the three archaic will types. Then will be farting about with the bits at the end. Then adding more unnecessary bits to make it nicer. Do the logic flows this weekend, try to do it all. Making it pretty will come after. You can play 14 hours of Civilization you can do this bullshit.