Add files via upload
You put the will date and will specific date above will type. Now take the switch statement out of the will specific date, instead assign a value to a variable based on the specific date, and have it feed to will type. Then you need to adjust the will type function to only accept answers that take into account the will specific date. And also adjust the show/hide at that point. That is the next step.