Add files via upload
This one you figured out that you couldn't display the next block of code because you'd put in the variable name of the show/hide button inside of the div you wanted to target instead of the div variable name itself. You coded in the logic for the will dates in relation to statute enactment, and the will types for signature at end of document requirement. Then you added a button for if there is a will signature on every page to address statutory and notarial will requirements, and put in the function for it and declared the variable, and you're getting 'ReferenceError: willSignatureEvery is not defined' right now and are blinded by immediacy.