Add files via upload
I noticed the Testator Clause button wasn't working. It turned out there were two functions with the same name. I spotted that and fixed it by changing the name on one, but then I didn't change the name of the function in both spots I needed to change it to. So I fixed the function name in the one spot and the Testator clause button works now.